LydiaAWS/src/main/java/com/amazonaws/lambda/demo/ListPlaylistSegmentsHandler.java
package com.amazonaws.lambda.demo;

import java.util.ArrayList;
import java.util.List;
import java.util.Scanner;

import com.amazonaws.regions.Regions;
import com.amazonaws.services.lambda.runtime.*;
import com.amazonaws.services.s3.AmazonS3;
import com.amazonaws.services.s3.AmazonS3ClientBuilder;
import com.amazonaws.services.s3.model.ListObjectsV2Request;
import com.amazonaws.services.s3.model.ListObjectsV2Result;
import com.amazonaws.services.s3.model.S3Object;
import com.amazonaws.services.s3.model.S3ObjectInputStream;
import com.amazonaws.services.s3.model.S3ObjectSummary;

import com.amazonaws.lambda.demo.db.SegmentsDAO;
import com.amazonaws.lambda.demo.model.Segment;
import com.amazonaws.lambda.demo.http.*;


public class ListPlaylistSegmentsHandler implements RequestHandler <ListPlaylistSegmentsRequest, ListPlaylistSegmentsResponse> {
	
	public LambdaLogger logger;
	
	/** Load from RDS, if it exists
	 * 
	 * @throws Exception
	 */
	List<Segment> getSegments(String playlistName) throws Exception {
		logger.log("in getSegments");
		SegmentsDAO dao = new SegmentsDAO();
		
		return dao.getAllSegmentsFromPlaylist(playlistName);
	}
	
	private AmazonS3 s3 = null;
	
	@Override
	public ListPlaylistSegmentsResponse handleRequest(ListPlaylistSegmentsRequest input, Context context) {
		logger = context.getLogger();
		logger.log("Loading Java Lambda handler to list all segments");
		
		ListPlaylistSegmentsResponse response;
		try {
			// get all segments
			List<Segment> list = getSegments(input.getName());
//			for (Segment s : systemSegments()) {
//				if (!list.contains(s)) {
//					list.add(s);
//				}
//			}
			response = new ListPlaylistSegmentsResponse(list, 200);
		} catch (Exception e) {
			response = new ListPlaylistSegmentsResponse(403, e.getMessage());
		}
		
		return response;
	}
	
}
